requirements.txt
black==25.12.0
pylint==4.0.4
pytest==9.0.2
PyYAML==6.0.3